agent/requirements.txt
httpx==0.27.2
pydantic==2.9.2
redis==5.2.0
async_timeout==4.0.2
pytest
pytest-asyncio
pyyaml